Refactored `rustcoalescence/algorithms/gillespie/src/gillespie/classical.rs`
