+++ rustcoalescence/algorithms/gillespie/src/gillespie/classical/initialiser/genesis.rs
@@ -1,0 +1,75 @@
+use necsim_core::{
+    cogs::{
+        EmigrationExit, ImmigrationEntry, LineageReference, LocallyCoherentLineageStore, MathsCore,
+        RngCore,
+    },
+    reporter::Reporter,
+};
+
+use necsim_impls_no_std::cogs::{
+    active_lineage_sampler::classical::ClassicalActiveLineageSampler,
+    dispersal_sampler::in_memory::alias::InMemoryAliasDispersalSampler,
+    origin_sampler::TrustedOriginSampler,
+};
+use necsim_partitioning_core::LocalPartition;
+
+use rustcoalescence_scenarios::Scenario;
+
+use super::ClassicalLineageStoreSampleInitialiser;
+
+#[allow(clippy::module_name_repetitions)]
+pub struct GenesisInitialiser;
+
+#[allow(clippy::type_complexity)]
+impl<M: MathsCore, G: RngCore<M>, O: Scenario<M, G>>
+    ClassicalLineageStoreSampleInitialiser<M, G, O, !> for GenesisInitialiser
+{
+    type ActiveLineageSampler<
+        R: LineageReference<M, O::Habitat>,
+        S: LocallyCoherentLineageStore<M, O::Habitat, R>,
+        X: EmigrationExit<M, O::Habitat, G, R, S>,
+        I: ImmigrationEntry<M>,
+    > = ClassicalActiveLineageSampler<
+        M,
+        O::Habitat,
+        G,
+        R,
+        S,
+        X,
+        Self::DispersalSampler,
+        O::SpeciationProbability,
+        I,
+    >;
+    type DispersalSampler = O::DispersalSampler<InMemoryAliasDispersalSampler<M, O::Habitat, G>>;
+
+    fn init<
+        'h,
+        T: TrustedOriginSampler<'h, M, Habitat = O::Habitat>,
+        R: LineageReference<M, O::Habitat>,
+        S: LocallyCoherentLineageStore<M, O::Habitat, R>,
+        X: EmigrationExit<M, O::Habitat, G, R, S>,
+        I: ImmigrationEntry<M>,
+        Q: Reporter,
+        P: LocalPartition<Q>,
+    >(
+        self,
+        origin_sampler: T,
+        dispersal_sampler: O::DispersalSampler<InMemoryAliasDispersalSampler<M, O::Habitat, G>>,
+        _local_partition: &mut P,
+    ) -> Result<
+        (
+            S,
+            Self::DispersalSampler,
+            Self::ActiveLineageSampler<R, S, X, I>,
+        ),
+        !,
+    >
+    where
+        O::Habitat: 'h,
+    {
+        let (lineage_store, active_lineage_sampler) =
+            ClassicalActiveLineageSampler::init_with_store(origin_sampler);
+
+        Ok((lineage_store, dispersal_sampler, active_lineage_sampler))
+    }
+}
