+++ rustcoalescence/algorithms/gillespie/src/gillespie/classical/mod.rs
@@ -1,0 +1,122 @@
+use necsim_core::{cogs::LocallyCoherentLineageStore, lineage::Lineage, reporter::Reporter};
+use necsim_core_bond::{NonNegativeF64, PositiveF64};
+
+use necsim_impls_no_std::cogs::{
+    lineage_reference::in_memory::InMemoryLineageReference,
+    lineage_store::coherent::locally::classical::ClassicalLineageStore,
+    maths::intrinsics::IntrinsicsMathsCore, origin_sampler::pre_sampler::OriginPreSampler,
+    turnover_rate::uniform::UniformTurnoverRate,
+};
+use necsim_impls_std::cogs::rng::pcg::Pcg;
+use necsim_partitioning_core::LocalPartition;
+
+use rustcoalescence_algorithms::{
+    result::{ResumeError, SimulationOutcome},
+    strategy::RestartFixUpStrategy,
+    Algorithm,
+};
+use rustcoalescence_scenarios::Scenario;
+
+use super::GillespieAlgorithm;
+
+mod initialiser;
+mod launch;
+
+use initialiser::{
+    fixup::FixUpInitialiser, genesis::GenesisInitialiser, resume::ResumeInitialiser,
+};
+
+// Optimised 'Classical' implementation for the `UniformTurnoverSampler`
+#[allow(clippy::type_complexity)]
+impl<
+        O: Scenario<
+            IntrinsicsMathsCore,
+            Pcg<IntrinsicsMathsCore>,
+            LineageReference = InMemoryLineageReference,
+            TurnoverRate = UniformTurnoverRate,
+        >,
+        R: Reporter,
+        P: LocalPartition<R>,
+    > Algorithm<O, R, P> for GillespieAlgorithm
+where
+    O::LineageStore<ClassicalLineageStore<IntrinsicsMathsCore, O::Habitat>>:
+        LocallyCoherentLineageStore<IntrinsicsMathsCore, O::Habitat, InMemoryLineageReference>,
+{
+    #[allow(clippy::too_many_lines)]
+    fn initialise_and_simulate<I: Iterator<Item = u64>>(
+        args: Self::Arguments,
+        rng: Self::Rng,
+        scenario: O,
+        pre_sampler: OriginPreSampler<Self::MathsCore, I>,
+        pause_before: Option<NonNegativeF64>,
+        local_partition: &mut P,
+    ) -> Result<SimulationOutcome<Self::MathsCore, Self::Rng>, Self::Error> {
+        launch::initialise_and_simulate(
+            args,
+            rng,
+            scenario,
+            pre_sampler,
+            pause_before,
+            local_partition,
+            GenesisInitialiser,
+        )
+    }
+
+    /// # Errors
+    ///
+    /// Returns a `ContinueError::Sample` if initialising the resuming
+    ///  simulation failed
+    fn resume_and_simulate<I: Iterator<Item = u64>, L: ExactSizeIterator<Item = Lineage>>(
+        args: Self::Arguments,
+        rng: Self::Rng,
+        scenario: O,
+        pre_sampler: OriginPreSampler<Self::MathsCore, I>,
+        lineages: L,
+        resume_after: Option<NonNegativeF64>,
+        pause_before: Option<NonNegativeF64>,
+        local_partition: &mut P,
+    ) -> Result<SimulationOutcome<Self::MathsCore, Self::Rng>, ResumeError<Self::Error>> {
+        launch::initialise_and_simulate(
+            args,
+            rng,
+            scenario,
+            pre_sampler,
+            pause_before,
+            local_partition,
+            ResumeInitialiser {
+                lineages,
+                resume_after,
+            },
+        )
+    }
+
+    /// # Errors
+    ///
+    /// Returns a `ContinueError<Self::Error>` if fixing up the restarting
+    ///  simulation (incl. running the algorithm) failed
+    #[allow(clippy::too_many_lines)]
+    fn fixup_for_restart<I: Iterator<Item = u64>, L: ExactSizeIterator<Item = Lineage>>(
+        args: Self::Arguments,
+        rng: Self::Rng,
+        scenario: O,
+        pre_sampler: OriginPreSampler<Self::MathsCore, I>,
+        lineages: L,
+        restart_at: PositiveF64,
+        fixup_strategy: RestartFixUpStrategy,
+        local_partition: &mut P,
+    ) -> Result<SimulationOutcome<Self::MathsCore, Self::Rng>, ResumeError<Self::Error>> {
+        launch::initialise_and_simulate(
+            args,
+            rng,
+            scenario,
+            pre_sampler,
+            Some(PositiveF64::max_after(restart_at.into(), restart_at.into()).into()),
+            local_partition,
+            FixUpInitialiser {
+                lineages,
+                restart_at,
+                fixup_strategy,
+            },
+        )
+    }
+}
